feat: added comments to utility functions and folder selector
