+++ src/extension.ts
@@ -2,11 +2,8 @@ import * as vscode from 'vscode'
 import * as path from 'path'
 import * as fs from 'fs'
 import { Menu } from './models/menu.model'
-import {
-  traverseOptions,
-  verifyTechnologyDependency,
-} from './utils/dependencies.utils'
-import { generateFunnyName } from './utils/app.utils'
+import { traverseOptions, verifyTechnologyDependency } from './utils/menu.utils'
+import { askForProjectName, selectProjectDirectory } from './utils/app.utils'
 
 export function activate(context: vscode.ExtensionContext) {
   const disposable = vscode.commands.registerCommand(
@@ -30,15 +27,15 @@ export function activate(context: vscode.ExtensionContext) {
       const command = await traverseOptions(technology.frameworks)
       if (!command) return
 
-      const funnyPlaceholder = generateFunnyName()
+      const projectName = await askForProjectName()
 
-      const projectName =
-        (await vscode.window.showInputBox({
-          prompt: 'Enter the name of your project',
-          placeHolder: funnyPlaceholder,
-        })) || funnyPlaceholder
+      const projectDirectory = await selectProjectDirectory()
+      if (!projectDirectory) return
 
-      const finalCommand = command.replace('${projectName}', projectName)
+      const finalCommand = `cd ${projectDirectory} && ${command.replace(
+        '${projectName}',
+        projectName,
+      )}`
       const terminal = vscode.window.createTerminal({
         name: 'Project Crafter Terminal',
       })
